feat(config): Split configuration into separate files
This commit splits the Neovim configuration into separate files for better organization
and maintainability. Configuration files for global settings, key mappings, and plugins
have been created.
