+++ nvim_config/global_settings.lua
@@ -1,0 +1,13 @@
+-- global settings
+vim.g.mapleader = ","
+local opt = vim.opt
+opt.number = true
+opt.relativenumber = true
+opt.hlsearch = true
+opt.incsearch = true
+opt.updatetime = 100
+opt.cursorline = true
+opt.showmode = false
+opt.shiftwidth = 2
+opt.smartindent = true
+vim.cmd('colorscheme gruvbox')
